refactor: Refactor action tracking functions to discrete module
Refs: #426
