+++ src/module/helpers/automation/combat.ts
@@ -1,45 +1,58 @@
 import { LancerCombat } from "lancer-initiative";
+import { modAction } from "../../action/actionTracker";
+import { LancerActor } from "../../actor/lancer-actor";
 import { prepareChargeMacro } from "../../macros";
-import { getAutomationOptions } from "../../settings";
+import { getActionTrackerOptions, getAutomationOptions } from "../../settings";
 
 export async function handleCombatUpdate(...[combat, changed]: Parameters<Hooks.UpdateDocument<typeof Combat>>) {
   //if (combat.round === 0 || changed?.round === 0) return;
   if (!("turn" in changed) && changed.round !== 1) return;
   if (game.combats!.get(combat.id!)?.data?.combatants.contents.length == 0) return;
 
-  const auto = getAutomationOptions();
-  if (auto.enabled) {
+  if (getAutomationOptions().enabled) {
     // TODO: Update foundryvtt typings.
     const nextActor = lookup(combat, (combat.current as any).combatantId);
     const prevActor = lookup(combat, (combat.previous as any).combatantId);
 
-    console.log(combat);
-
     // Handle refreshing for next combatant.
     if (nextActor) {
-      console.log(`Processing combat automation for ${nextActor.name}`);
-
-      // Handle NPC charges.
-      prepareChargeMacro(nextActor);
-
-      // Refresh actions.
-      console.log(`Next up! Refreshing [${nextActor.data.name}]!`);
-      game.action_manager?.modAction(nextActor, false);
-
-      // Refresh reactions on new turn.
-      refreshReactions(combat);
+      processStartTurn(nextActor);
     }
 
     // Handle end-of-turn for previous combatant.
     if (prevActor) {
-      // Dump extra actions.
-      console.log(
-        `Turn over! [${prevActor.data.name}] ended turn with ${JSON.stringify(
-          prevActor.data.data.action_tracker
-        )}`
-      );
-      game.action_manager?.modAction(prevActor, true);
+      processEndTurn(prevActor);
     }
+
+  }
+}
+
+function processStartTurn(actor: LancerActor) {
+  console.log(`Processing start-of-turn combat automation for ${actor.name}`);
+
+  // Handle NPC charges.
+  prepareChargeMacro(actor);
+
+  // Refresh actions.
+  modAction(actor, false);
+
+  // Refresh reactions on new turn.
+  refreshReactions(game.combat);
+
+  // Print chat messages.
+  if (getActionTrackerOptions().printMessages) {
+
+  }
+}
+
+function processEndTurn(actor: LancerActor) {
+  console.log(`Processing end-of-turn combat automation for ${actor.name}`);
+
+  // Dump extra actions.
+  modAction(actor, true);
+
+  // Print chat messages.
+  if (getActionTrackerOptions().printMessages) {
 
   }
 }
@@ -50,10 +63,12 @@ function lookup(combat: LancerCombat, id: String | null) {
   } else return undefined;
 }
 
-function refreshReactions(combat: LancerCombat) {
-  combat.data.combatants.map(comb => {
-    if (comb.actor) {
-      game.action_manager?.modAction(comb.actor, false, "reaction");
-    }
-  })
+function refreshReactions(combat: LancerCombat | null) {
+  if (combat) {
+    combat.data.combatants.map(comb => {
+      if (comb.actor) {
+        modAction(comb.actor, false, "reaction");
+      }
+    })
+  }
 }
